Fix exception thrown for custom column in SQLite (#125)
This update fixes an exception that was thrown when using a custom
column name in SQLite. When using a custom column name, the custom name
was used as the parameter name. Upon setting the parameter, however, it
tried to set the parameter with the property name instead. This commit
fixes this error by using the property name in both places.